refactor: renames testmock to mock
