+++ http/mock/request.go
@@ -1,4 +1,4 @@
-package mocks
+package mock
 
 import (
 	"net/http"
@@ -8,7 +8,7 @@ import (
 	"go.keploy.io/server/pkg/models"
 )
 
-type MocksReq struct {
+type MockReq struct {
 	ID       string                            `json:"id" bson:"_id"`
 	Created  int64                             `json:"created" bson:"created"`
 	Updated  int64                             `json:"updated" bson:"updated"`
@@ -17,7 +17,7 @@ type MocksReq struct {
 	Deps     []map[string]models.FetchResponse `json:"deps" bson:"deps,omitempty"`
 }
 
-func (req *MocksReq) Bind(r *http.Request) error {
+func (req *MockReq) Bind(r *http.Request) error {
 	if req.ID == "" {
 		req.ID = uuid.New().String()
 	}
