+++ pkg/service/mock/mock.go
@@ -1,4 +1,4 @@
-package mocks
+package mock
 
 import (
 	"context"
@@ -7,25 +7,25 @@ import (
 	"go.uber.org/zap"
 )
 
-func NewMockService(c models.MockDB, log *zap.Logger) *Mocks {
-	return &Mocks{
+func NewMockService(c models.MockDB, log *zap.Logger) *Mock {
+	return &Mock{
 		sdb: c,
 		log: log,
 	}
 }
 
-type Mocks struct {
+type Mock struct {
 	sdb models.MockDB
 	log *zap.Logger
 }
 
-func (s *Mocks) Insert(ctx context.Context, doc models.Mock) error {
+func (s *Mock) Put(ctx context.Context, doc models.Mock) error {
 	if count, err := s.sdb.CountDocs(ctx, doc.AppID, doc.TestName); err == nil && count > 0 {
 		return s.sdb.UpdateArr(ctx, doc.AppID, doc.TestName, doc)
 	}
-	return s.sdb.Insert(ctx, doc)
+	return s.sdb.Put(ctx, doc)
 }
 
-func (s *Mocks) Get(ctx context.Context, app string, testName string) ([]models.Mock, error) {
+func (s *Mock) Get(ctx context.Context, app string, testName string) ([]models.Mock, error) {
 	return s.sdb.Get(ctx, app, testName)
 }
