+++ server/server.go
@@ -21,11 +21,11 @@ import (
 	"github.com/keploy/go-sdk/keploy"
 	"go.keploy.io/server/graph"
 	"go.keploy.io/server/graph/generated"
-	"go.keploy.io/server/http/mocks"
+	"go.keploy.io/server/http/mock"
 	"go.keploy.io/server/http/regression"
 	"go.keploy.io/server/pkg/platform/mgo"
 	"go.keploy.io/server/pkg/platform/telemetry"
-	mocks2 "go.keploy.io/server/pkg/service/mocks"
+	mock2 "go.keploy.io/server/pkg/service/mock"
 	regression2 "go.keploy.io/server/pkg/service/regression"
 	"go.keploy.io/server/pkg/service/run"
 	"go.keploy.io/server/web"
@@ -73,7 +73,7 @@ func Server() *chi.Mux {
 	rdb := mgo.NewRun(kmongo.NewCollection(db.Collection(conf.TestRunTable)), kmongo.NewCollection(db.Collection(conf.TestTable)), logger)
 
 	mdb := mgo.NewMockDB(kmongo.NewCollection(db.Collection("test-mocks")), logger)
-	mockSrv := mocks2.NewMockService(mdb, logger)
+	mockSrv := mock2.NewMockService(mdb, logger)
 	enabled := conf.EnableTelemetry
 	analyticsConfig := telemetry.NewTelemetry(mgo.NewTelemetryDB(db, conf.TelemetryTable, enabled, logger), enabled, keploy.GetMode() == keploy.MODE_OFF, logger)
 
@@ -124,7 +124,7 @@ func Server() *chi.Mux {
 	// add api routes
 	r.Route("/api", func(r chi.Router) {
 		regression.New(r, logger, regSrv, runSrv)
-		mocks.New(r, logger, mockSrv)
+		mock.New(r, logger, mockSrv)
 
 		r.Handle("/", playground.Handler("keploy graphql backend", "/api/query"))
 		r.Handle("/query", srv)
